Revert "Prevent combing from going over internal holes"
This reverts commit aa77cddd5a8b805f315eb2e724557686bbc2aad6.
That commit was making the wrong assumption that holes are not visible from the outside. They are. And there might be other shapes inside of the hole, such as nested rings. Internal holes can have any shape, and don't necessarily need to be convex. They can have protruding parts that need to be avoided. We need to treat inside walls just the same way as outside walls.

Contributes to issue CURA-8117.