Fix bug: Always support previous branch.
Since the previous grounding location is never taken into account if the root is either removed (if outside of the current outline), or no longer a root (if inside) (due to slopes), the new grounding location found for the reconnected root could be far from the root of the previous layer. This caused the 'same' branch the next layer down (from root to first child) to no longer overlap with the higher layer. This is fixed by remembering the previous grounding location, and, if it is indeed there, to use it (taking into account the current outlines of course, that's why the intersection of the polygon with the line defined by the child and the old ground location is sought.)

part of CURA-8550