Prevent re-attach of a tree to go through itself.
If the attachment of a tree to another tree or a grounding location (on the border of a polygon) simply reroot the to-be-attached tree so that it doesn't intersect when the re-attach connection is made (to the grounding location). Turns out we didn't have the reroot functionality yet, so Ihad to write that as well.

part of finishing up CURA-8550 and the last remaining issue in bug CURA-8616 ... well hopefully at least