Don't attach to new trees if the wall is close.
There already was a small (magic number) value, but we actually _have_ a wall support radius already, so that should have been used in the first place. Bonus is removal of magic :-) This actually mitigates but not _solves_ a (large?) part of 'bug 1' and 'bug 2' from A's list (which was also reported by G as 'bug B' on github in the initial PR that was merged).

part of finishing up CURA-8550