Properly fix 'grounding location' issues when rerooting trees.
Tree from the previous (upper) layer get disconnected when the outlines aren't the same in this layer. That can happen due to various reasons, and can cause various issues. If the outlines become more spacious, then the node that was previously attached to a wall should now make a new connection to a wall ... that's the easy bit. If the trees from the above layer, that where propagated down, are now intersected by (outline) wall, they should still generally follow the same path back towards the rim that thir 'parents' did. If the nodes that are still inside while their parent node was outside (so they're now, perhaps only until rerooting is done, new roots) where to be 'normally' reconnected, then, while from the perspective of the current layer itself the tree would be correct, but it wouldn't nescesarily support their parent.
This was what the grounding location was originally used for already, but more as a sort of heuristic to mitigate the problem. This commit actually solves that completely, or near enough that I can't tell the difference.

part of CURA-8550