Fix branches intersecting models again.
Some definsive programming (more strict 'but it has to be the actual intersection of the line _segments_'). Mostly though remove some old code that's now been superseded by the 'last ground location' that was causing some trouble all of a sudden, after the previous fixes.

part of CURA-8550 (I guess this is also 'bug 2' as reported by A. in the Jira ticket which turns out to have been different from 'bug 1' after all.)