Use proper coordinate types for coordinates when extruding
This would theoretically increase memory usage, but in practice leads to compiler optimisations since it won't need to cast to smaller ints. Let's hope it's not a noticeable impact.
This also seems to fix a rare bug where the references to these coordinates were invalidated when exporting g-code. Ints and coord_t shouldn't be passed by reference anyway.

Contributes to issue CURA-8622.