Don't calculate intersections of lines that are really far away
This can lead to overflow errors. If we start using points with those really far coordinates, we may end up with coordinates where we will get integer overflows if we calculate distances between them.
The coordinate check performed here checks if the points are within 2.1km of the origin. I think that's pretty safe for most printers. When we get printers to print cities we'll be in trouble, but we'll see about that when we get there...

Such integer overflows were causing the simplify function to accept intersection points that were extremely far away, since the calculated distance to them ended up negative. These far intersection points were then included in the resulting polygon, resulting in attempts to fill polygons of literally 300km length. Not nice.

Contributes to issue CURA-8622.