Apply 10um closing operation to guard against rounding errors
Nobody wants to print infill lines of 10 micron, really. Not intentionally, anyway.
There is a rounding error here where the original border area and the skin area don't exactly match up sometimes. They should match up, but if there is a long diagonal of which only part is bordering skin, there is often a 0-2um border between them. This border gets exploded a lot when the infill overlap setting is applied. A closing operation before that prevents this effect. And because the expansion of the closing operation can be taken along with the infill overlap it costs only 1 extra offset, not 2.

Contributes to issue CURA-8644.