Make sure infill knows about perimeter gaps.
Otherwise, areas that are large enough to be not deleted, but thinner than an entire wall, will be filled with both infill and perimeter gaps... and other similar stuff that was caused by not taking the literal infill area, but dome sort of derived area. What is done now instead is taking that into acount when _making_ the infill area.

part of CURA-8644