Don't fill the entire outline when processing each part.
Ligning infill was first generated, then applied to the part. However, the enitre fill would be applied to each part. If a model consisted of more than one part, the lightning paths would be retraced for each part of the model in that layer. Now it limits the output to the outlines that are currently processed. It's relatively efficient, since it only has to check one node of each tree against the outlines of a single part.

fixes Cura/#10793