Move inside by 1 micron to prevent edge cases
When moving inside by 0 microns, the point is placed on the border. This would be fine if border_is_inside was working correctly, but unfortunately due to rounding errors it is still often considered to be outside. Adding 1 micron doesn't hurt, and resolves the edge case.
However it's still not working properly due to the relation between X, Y and Z distance. But that's a separate matter that I'll try solving next.

Contributes to issue CURA-8635.