Break long line segments in pieces if deviation is exceeded
This breaks a long line into multiple pieces if the Maximum Extrusion Area Deviation is exceeded. Long line segments weren't broken up by the Skeletal Trapezoidation if there was no additional rib in the skeletal graph. There would only be additional ribs if the number of lines changes, or if there is a vertex in the contour. This fixes that issue: It will now test if a line needs to be broken up into multiple pieces, by evaluating how much a line deviates in its area from the maximum deviation.

If it is exceeded, it will calculate how many pieces it needs to use to stay beneath the maximum extrusion area deviation. If this causes the pieces to be shorter than the maximum resolution, the maximum resolution is held. Thus this implements the following constraints:
* Lines should not get broken up into pieces smaller than the Maximum Resolution.
* If possible with the first constraint, lines should get broken up into the fewest amount of pieces needed to hold the Maximum Extrusion Area Deviation.

Each piece gets its line width calculated from the average width along its length.

Implements CURA-8170.