Fix compilation warning binding to a temporary
It's not a problem here because we only use the temporary for as long as it exists, but it's good style to make the reference explicit and have the correct type.
Can't use ConstPolygonRef here because that constructs ConstPolygonRef in a temporary. Instead, let's just send its base class which is what we get there anyway due to ClipperLib::paths::const_iterator pointing to those.