Add option for removeDegenerateVerts to work with polylines
In this case, it won't compare the two endpoints of the polyline to see if they are in line with the previous/next vertices. It should never remove the first and last vertex of a polyline, even if they are in line. With a polygon this would be degenerate since the line would loop back on itself, but with a polyline it doesn't actually loop, so it doesn't loop back on itself either.

Contributes to issue CURA-8903.