fix inner contour not being generated on all sides
Sometimes the inner contour is very close to the middle.
If only one side of the contour gets generated then a hollow cylinder might get filled,
because the inner contour of only one side of the circle is generated.

With this commit the inner contours always overlap with the walls by 5 micron,
which should be negligible.