Indicate that extruders are used when raft forces them to prime
The raft explicitly forces all used extruders to be primed on the first layer (which is then a negative layer index, the raft base layer). However when doing so, it would try to prime, but then discover that the correct moment to prime is a few layers higher. So it wouldn't prime, and the prime would be skipped as a whole if it happens in a higher raft layer.
That the higher raft layers skip the prime as a whole is a bug too. However even if that is fixed, it would cause the prime to occur during the higher raft layers instead of down here at the first raft layer where it is supposed to.

Contributes to issue CURA-8960.