Re-enable basics of fuzzy skin.
At the moment, just replaced the polygons with their nearest equivalent of extrusion-lines, also still kept the skin-outside-only option as disabled. This has some issues that need to be resolved (seams, proper line-width handling, perimeter-nr. and the aforementioned disable for holes) but seems promising. There might also have been a fix to 4.13.x that we want to merge back into here, find out if it's already in here or needs to be backported ... or, well, forward-ported I suppose.

part of CURA-7887