Update interface of PolygonConnector to accept variable-width paths too
It works a little bit different now. Because there are two types of output, I'm giving it output parameters. And there's two types of input as well now.
This change effectively makes it remove concentric patterns now when connecting it. Because the concentric patterns are variable-width paths and those are not yet returned in the output.

Contributes to issue CURA-7828.