Move connect() entrypoint into a generic subroutine
The subroutine is meant to become generic. It accepts a template parameter (which is why it's in the header) and does some of it generically. However it calls more subroutines that now will need to become either generic (to reduce code duplication) or specialised (to make it work for each data type). If it were used on the VariableWidthPaths it would now not compile, since the subroutines can't be ran on that yet.

Contributes to issue CURA-7828.