Make getBridge generic and simplify
Our findConnection already returns failure if either of the connections couldn't be made, so we can simply check that once. It already returns both or neither.
We still need to check if they need to be swapped.
This function can now deal with either polygons or variable-width paths.

Contributes to issue CURA-7828.