Split outer and inner wall widths again (remove weighted average).
In the situation as it is now, (weighted) averaging the outer- and inner wall bead widths didn't make much sense. Several things have changed to make this true, but one of the most important ones is that the outer wall is now also kept the same widths as much as possible. (The only variation allowed is within a space that has a width of less than two walls). In the mean time, several things have been build on top of each other, resulting in the Redistributed strategy as it is now. Since we now have a much clearer image (literally in some cases) what the expected case should be, we can both throw out the now-nonsense weighted average, and both simplify and make right Redistributed.

In other words, since the original problem was what to do with separate values for outer and inner wall widths, but now we _also_ decided to keep the outer walls 'steady' -- instead of having this complicated system, just solve problem 1 with problem 2.

part of CURA-9027