Fix using wrong junctions when calculating areas.
The code as it was assumed that each junction specified the width for the _previous_ line it was connected to, but it actually specifies the width for the _next_ line it's connected to. This caused an issue where it could make the symetrical beadings so far out of kilter that it's easily visible in the preview. Technically this still can happen, but only because the Maximum Extrusion Area Deviation is set to a very large value in the front-end, so the simplification doesn't really care about the weights. We should probably put it back to 2000 square micron, as it was.

part of CURA-9036 (should fix it in fact)