Fix comb max moveInside distance multiplied for each object
This code is calculating the innermost wall line width. To do this, it applies the initial layer line width factor if it's the first layer. However instead of multiplying the calculated line width, it was multiplying the maximum line width that was found so far. As a result, if you had 10 objects, this max_inner_wall_width was multiplied by the initial layer line width factor 10 times, resulting in 0.8*1.2^10=5.0mm instead. And this got worse and worse if there are more meshes on the build plate.

Contributes to issue CURA-8131.