Add missing parameter to infill constructor
A bit of an insidious issue: The Infill constructor has so many parameters that you can't tell when one of them is missing. And some of the parameters are optional, so the compiler won't warn you about it. It even automatically casts between the bool and int type parameters so it's completely transparent except that the parameters get the wrong value when e.g. an enum has the value 0 instead of 1.

Recently we added another new parameter to this constructor: skip_line_stitching, added by 08d0354fd1a6eefd1000d37da88fd37430c95221. This parameter was set here and there where it needed to depend on whether the pattern was a monotonically printed line pattern. But other uses of this constructor were not modified. And so the parameters after that were used to determine whether it needed to skip_line_stitching. Those parameters often had completely unrelated values, in particular whether zigzags need to be connected. Because zigzags are connected in some cases, it would in those cases skip the stitching instead. And all other parameters would also shift by 1 leading to very weird results.

Fixes issue CURA-9077.