Fixing CICT
A container is no longer needed. We can build
directly on the VM since it is the same ubuntu
version as our container.

Build script will first install conan
and then pull the repositories.
Including arcus for which I added a Conan
package to the ultimaker remote artifactory.

These dependencies will in all likelyhood be
recompiled at each run. Normally you would
upload the build conan artifacts but the
current artifactory server is still a test
server.

I also still need to fix the Conan target
naming for the GTest.

Contributes to CURA-8640